Revert "Organise code"
This reverts commit 7fef02c377c43c4f3192eceb30862a1fd0431234.
